Add minimal version of SHAP explanations using KernelExplainer
This commit adds a `shap_explain` method to the `AutoGluonImputer`
- class. This is a way of interpreting black box models that scales
exponentially with number of columns when no subsampling takes place.

I also added a force plot to the sample notebook.

The implementation includes a TODO, as I didn't properly solve the
issues that arise when you pass a baseline to the explainer object
that is returned by `shap.KernelExplainer`.