Add saving and loading
Added saving and loading capabilities to AutoGluonImputer. This was
only tested on Fedora 34 and might run into problems on MacOS or
Windows.

It feels weird that Autogluon saved automatically to the disc, while
AutoGluonImputer's `save` method has to be called to save the object's
attributes. One might consider streamlining this behavior.