[wip] start implementing test
We're experiencing some unexpected behavior regarding the way we
determine thresholds in `.fit()`. Which is why I've started looking into
adding some tests to the code.

The first test is a very basic integration test. It is supposed to
ensure that the core feature of Autowig, thresholding and imputing
values with that, works.

I got inspired by the integration tests written for the old
`SimpleImputer`. The current state of the branch doesn't properly work.