Add two integration tests
There are now two integration tests. Both currently fail, because
`imputer.predict()` returns missing values. Missing values are returned
because a class probability is below the computed precision threshold.

- `test_precision_threshold` tests the ability to impute categorical
data. The precision threshold is set to `0.01`, which makes me believe
that this hints at a bug.

- `test_quantile_regression` tests the ability to impute numerical
data. The precision threshold is set to `0.01`, which also makes me
believe that this hints at a bug.