Refine Tests
There are now two tests in place: An integration test for quantile
regression, and a test for precision thresholds in classification tasks.

The test for quantile regression is just a scaffolding for now, as the
API is still a bit weird and I expect it to change in the future -- we
compute the quantile-regresssion thresholds during `.fit()`, but the
classification thresholds during `.predict()`. I find that odd.

I added an assertion to `test_autogluon_imputer_precision_threshold`
that ensures that empirically measured precision thresholds increase
monotonously with increased user-specified precision thresholds.